Move ComScoreTracker
